Create initial project
I am writing a new Django application.  The goal of the software is to allow a large team to manage a glossary of words.  What is dinstinctive is that the words can have different meanings in different (mostly technical) domains.  There will be an approval process as well so people can propose new definitions and other two others must approve before the new words or definitions are adopted.  First steps first, we'll start working on the Django admin (not any UI yet).  Let's first set up a new Django project in this directory for the application which we'll call Termageddon.

My admin/admin credentials did not work.