rich text, approval flow, test passwords
Turn the definition field into a rich-text field and update the application accordingly.  Update the application so that the front-end interface supports proposing edits to definitions and also supports the approval process.  In addition, introduce pagination and filtering to support a large number of terms.  This is early development, so there is no need to support migrations.  Instead, after each round of edits which changes the data model, reset the database using the Django management command we created.  Create a root-level README.md file with with development instructions.  It should include a link to the Django documentation and a brief description of the database reset management command.  Update the database reset command so that, when loading from a CSV, the created test users all have a password matching their last name.