pills; infinite scroll; login/logout
When viewing terms in the main view, should pills indicating the domains for which definitions are available.  Also, in that main view, there should be infinite scrolling which loads the next page of terms as you scroll.  Along the top header througout the frontend (not the Django admin), we need an indication of who is the logged in user and ability to log out and log in.