2 approvers
Items should remain on the approval board until they've gotten the minimum number of approvals (we'll start with 2), none of whom is the author.  The approval board should indicate for each proposed term who has approved it so far.  If a person is the author or has already approved, the Approve button should be disabled.