Misc layout changes
Let's tighten up the layout a bit:  (1) Move the pills to the right of the terms.  (2) The logged in user should be full name, and without the Welcome greeting.  (3) The Review should be a clearer button in the top header --- it's now black on black; lighten it up so it pops out.  (4) there should be a button "Glossary" which returns to the list of terms.  I know this is redundant to the home button Termageddon, but it's clearer.  (5). There should be an button "Admin" to go to the Django Admin which is enabled only for users who can access the Django Admin.  (6). No page should be viewable unless the user is authenticated.  If the user is not authenticated, display a splash login page.