Fix alphabetization of diacrytics
I noticed that the word ångström(Å) is alphabetized after z rather than with A.  I don't know that this is worth addressing, since it's not likely to happen in practice.  But is there a simple way to have more robust Alphabetization taking into account such special characters that people expect to see alongside the English letters?

The makemigrations is trying to figure out how to handle pre-existing data in the database.  Let's instead just remove the database, remove all existing migrations, and re-write a base migration.  We have no existing data, so there is no need to worry about migrations.