First try at links
OK --- Definitions might include external and internal links.  The external links should be valid URLs and should be pre-validated to return a 200 GET response.  The internal links will be to Definitions, so there will need to be a pop-up selection window with search to select an existing Term and then the appropriate Definition of that Term.