External links
OK -- not quite what I had in mind.  Instead, I'm hoping for two different tinyMCE icons.  One for a generic link, which only needs a URL and text to display.  But it should always open in another window, and isn't link to a term.  So nix the bottom two dropdowns.  Once that is done we can work on a separate selection for links to Definitions.