Links
OK --- let's go with your suggestion for the icon.  It's easy to change it later if I find something better.  Let's put the icon to the right of the link icon in the tinyMCE editor.  When opened, it brings up a popup which allows you to first search and select a Term.  Once a Term is selected, you can search and select a Domain which has a Definition.  What is stored is the Term/Domain pair.  When clicked, it opens an approved Definition with that Term/Domain pair.  I think that it's acceptable for it to open in the same window, since it's still withing the application.