TinyMCE continued
Still an error, "Could not find a valid required value for domain_id in..."  In any event, let's replace the dropdown select for "Select Term" with an infinite scrollable text box that has room for about 10 terms and is the width of about 30 characters.  Then it's easier to see the response of the terms listed in that text box when you type in the search box.  The Domain selection should default to unselected so the user is forced to select a Domain.  The definition should appear below once a Domain is selected so the user can confirm they picked what they wanted.