tinyMCE continued
Better.  No errors, and the pre-select is coming into shape.  Now the pre-selected words should show in a box which initializes to the first page of all words.  Currently there is now way to "select" a word which should highlight it.  (The words are not currently clickable.)  Once selected, then the Domain selection logic you've put in place should work.