tinyMCE continued
There are still issues.  On loading a term page, I get the error "theme.min.js:1 Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "0 0 24".  I don't know if I should worry about it.  But in the tinyMCE popop for "Link to Definition", the Search for a term box resets after the debouce, but the first page of words which it should search for based on the search text properly loads.  But there is still no infinite scroll to allow to get the later matching words, nor are the words clickable to select and highlight.
@Angular